Solved a little bug with buffer access

+++ app/src/main/cpp/vectorSorter.h
@@ -5,6 +5,6 @@
 #ifndef DROIDPAPEREVOLUTION_VECTORSORTER_H
 #define DROIDPAPEREVOLUTION_VECTORSORTER_H
 
-void setBorders(int windowSize, int width, int offScreen, float *vector, float height, float roughness, float scale, float seed = -2);
+void setBorders(int windowSize, int width, float offScreen, float *vector, float height, float roughness, float scale, float seed = -2);
 
 #endif //DROIDPAPEREVOLUTION_VECTORSORTER_H
